Fix unescaped element contents
`DOMDocument::createElement()` has a ~~bug~~ documented oddity in that
while one would expect the second parameter, `$contents`, to be text-safe,
it is actually supposed to be used for parsed character data. If you
wish to store a string with an ampersand in it, you'll either have to
manually call `htmlspecialchars()` or use an 8-line snippet that creates
a `DOMTextNode` each time.

This patch creates a helper function for option #2.